Add files via upload
I originally created a repository with a zip of all of these (img folder, index, & style) and a read me that consisted of text describing the project. When I deployed, it showed the Readme and nothing else. I completely deleted that repository and created a new one without a Readme. I uploaded the zip again and upon deployment, github remembered my old Readme file without it present within the new repository. After a couple quick Googles to learn that github does not intuitively extract zipped folders and that I could drag and drop the img folder into uploads rather than clicking to upload files (which would not allow for a full file upload), I think I have the content necessary to deploy the site.