Add files via upload
Attempting to have AI build my website for me was a huge mistake. I will need to do more research and experimenting to achieve the full look that I have in mind. For now, this is legible and all images upload and with that, I am not mad at it.